implement draggable windows
The visualiser now comes with two windows, each of which can display an
interactive egg box diagram. However the console and some other features
are missing. Ultimately the plan is to allow the user to view a quotient
semigroup alongside the parent semigroup and the parent's congruence
lattice. However that won't be possible until I implement a graph
drawing method for drawing the congruence lattice and non-linear eggbox
diagrams.